Refactor architecture (#16)
* contracts: add minter

* contracts: remove collection, add registry

* dep: update solady

* contracts: misc

+++ test/utils/Constants.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.25;
 abstract contract Constants {
     uint32 constant TEST_TOKEN_MAX_SUPPLY = type(uint32).max;
     uint128 constant TEST_TOKEN_PRICE = 0.05 ether;
-    uint32 constant TEST_TOKEN_MINT_DURATION = 365 days;
+    uint40 constant TEST_TOKEN_MINT_DURATION = 365 days;
     string constant TEST_TOKEN_URI = "https://token.com/uri";
     string constant TEST_COLLECTION_NAME = "Test Collection";
     string constant TEST_COLLECTION_SYMBOL = "TST";
